+++ database/migrations/2024_09_25_111856_create_factures_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('factures', function (Blueprint $table) {
             $table->id();
             $table->dateTime('DateFact'); 
-            $table->unsignedBigInteger('idClient'); // Clé étrangère vers clients.id
+            $table->Integer('idClient'); // Clé étrangère vers clients.id
             $table->timestamps();
 
             $table->foreign('idClient')
